+++ tests/LiquidBox/Tests/Plates/Extension/RoutingTest.php
@@ -2,7 +2,11 @@
 
 namespace LiquidBox\Tests\Plates\Extension;
 
+use LiquidBox\Silex\Provider\PlatesServiceProvider;
+use Silex\Application;
 use Silex\WebTestCase;
+use Symfony\Component\HttpFoundation\Request;
+use Symfony\Component\Routing\Generator\UrlGenerator;
 
 /**
  * @author Jonathan-Paul Marois <jonathanpaul.marois@gmail.com>
@@ -11,25 +15,31 @@ class RoutingTest extends WebTestCase
 {
     public function createApplication()
     {
-        $app = new \Silex\Application();
+        $app = new Application();
 
-        $app->register(new \LiquidBox\Silex\Provider\PlatesServiceProvider(), [
+        $app->register(new PlatesServiceProvider(), [
             'plates.directory' => __DIR__ . '/../../Resources/views',
         ]);
-        $app->register(new \Silex\Provider\UrlGeneratorServiceProvider());
 
-        $app->get('/', function () {})->bind('homepage');
-        $app->get('/hello/{name}', function ($name) {})->bind('hello');
+        $app['url_generator'] = function ($app) {
+            $app->flush();
+            return new UrlGenerator($app['routes'], $app['request_context']);
+        };
 
-        $app['request'] = \Symfony\Component\HttpFoundation\Request::createFromGlobals();
+        $app->get('/', function () {
+        })->bind('homepage');
+        $app->get('/hello/{name}', function ($name) {
+        })->bind('hello');
+
+        $app['request_stack']->push(Request::createFromGlobals());
 
         return $app;
     }
 
     public function testRegister()
     {
-        $this->assertTrue($this->app['plates']->doesFunctionExist('path'));
-        $this->assertTrue($this->app['plates']->doesFunctionExist('url'));
+        $this->assertTrue($this->app['plates']->doesFunctionExist('path'), 'Function path() does not exist.');
+        $this->assertTrue($this->app['plates']->doesFunctionExist('url'), 'Function url() does not exist.');
     }
 
     public function testPathCall()