+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2017-07-31
+
+### Added
+
+* Support for Silex 2.x.
+
+### Removed
+
+* Support for package rych/silex-plates-provider and alias plates.path of parameter plates.directory.
+* PHPUnit tests for PHP 5.4.
+
 ## 1.0.0 - 2017-07-03
 
 ### Added
@@ -12,3 +23,5 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * Support for [Plates](https://packagist.org/packages/league/plates) 3.x.
 * Support for [PSR2 coding standard](http://www.php-fig.org/psr/psr-2) and [PSR4 autoloading](http://www.php-fig.org/psr/psr-4).
 * Support for PHPUnit tests with [Travis CI](https://travis-ci.org) and [Scrutinizer CI](https://scrutinizer-ci.com).
+
+[2.0.0]: http://github.com/liquidbox/silex-plates/compare/v1.0.0...v2.0.0